refactor(digitsd): extract voicemail callbacks into voicemail.go
Pure file-split refactor. No behavior change.

main.go shed 1069 lines: all voicemail-prefixed phone.Callbacks methods
(VoicemailPickup, VoicemailAutoAnswer, VoicemailRecordEnded,
VoicemailEnabled, VoicemailRecordGreeting,
VoicemailRecordGreetingKey, VoicemailDeleteGreeting,
VoicemailRetrievalCode, VoicemailEnterPlayback, VoicemailExitPlayback,
VoicemailKey), the playback machinery
(voicemailPlaybackSession, voicemailPlaybackLoop,
voicemailAdvanceFromEOF, openNextUnheardLocked, reopenLocked,
teardownPlaybackLocked, closeMixerSourceLocked,
voicemailExitToDialtoneAsync), the greeting playback helpers
(playVoicemailGreeting, playCustomGreeting, playDefaultGreeting,
finalizeGreetingRecording, greetingPlaybackDuration), and the LED
helpers (evaluateLED, ledModeWithVoicemailHint).

voicemail.go picks up the same code under package main with the same
receiver (*daemonCallbacks). The daemonCallbacks struct definition and
its voicemail-related fields stay on main.go: only methods moved.
Imports trimmed: main.go no longer needs errors (the io / context /
fmt / etc. imports it kept all still have call sites outside the
voicemail block).

Verification: go build ./..., go test ./... (including
TestLEDModeWithVoicemailHint), golangci-lint run ./..., make build
(Docker arm64) all clean. Net delta +25 lines covers the new package
doc, import block, and spacing.
